refactor: extract chart generation into ChartService and deduplicate RedisConfig
